rename service folder to be /mentorship/contracts
