+++ src/pages/student/dashboard/mentorships/mentorship-contract/index.js
@@ -16,7 +16,7 @@ import { isNil } from 'lodash'
 import React, { useEffect, useState } from 'react'
 import { Helmet } from 'react-helmet'
 import { useParams } from 'react-router-dom'
-import { getContract, terminateMentorshipContract } from 'services/mentorship/subscription'
+import { getContract, terminateMentorshipContract } from 'services/mentorship/contracts'
 import { requestRefund } from 'services/wallet'
 
 const MentorshipContractView = () => {
